Create deletme.html
This program utilizes panel data on macroeconomic factors for 10 OECD countries recorded at an annual frequency from 1960 to 2013. The data is used to generate a table of stylized facts that highlight the correlation of international real business cycles.

MATLAB is used to tabulate the values, which are then exported to Excel for tabular formatting. 

%% 
% Sources for the data are listed as follows
%       OECD: stats.oecd.org
%       TED: https://www.conference-board.org/data/economydatabase/
%       Penn World Tables Version 9.0: www.ggdc.net/pwt